+++ src/index.spec.ts
@@ -1,0 +1,12 @@
+/**
+ * indexテスト
+ */
+import * as assert from 'assert';
+
+import * as domain from './index';
+
+describe('import domain', () => {
+    it('ドメインをオブジェクトとしてインポートできるはず', () => {
+        assert.equal(typeof domain, 'object');
+    });
+});