+++ src/repo/invoice.ts
@@ -1,7 +1,0 @@
-import { repository } from '@cinerino/domain';
-
-/**
- * 請求書リポジトリ
- */
-export class MongoRepository extends repository.Invoice {
-}