+++ src/repo/paymentMethod.ts
@@ -1,7 +1,0 @@
-import { repository } from '@cinerino/domain';
-
-/**
- * 決済方法リポジトリー
- */
-export class MongoRepository extends repository.PaymentMethod {
-}