+++ src/repo/person.ts
@@ -1,7 +1,0 @@
-import { repository } from '@cinerino/domain';
-
-/**
- * 会員リポジトリ
- */
-export class CognitoRepository extends repository.Person {
-}