+++ src/repo/seller.ts
@@ -1,7 +1,0 @@
-import { repository } from '@cinerino/domain';
-
-/**
- * 販売者リポジトリ
- */
-export class MongoRepository extends repository.Seller {
-}