+++ src/repo/task.ts
@@ -1,7 +1,0 @@
-import { repository } from '@cinerino/domain';
-
-/**
- * タスクリポジトリ
- */
-export class MongoRepository extends repository.Task {
-}