+++ CHANGELOG.md
@@ -18,6 +18,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v31.0.0 - 2019-03-08
+
+### Changed
+
+- repositoryを完全にCinerino化
+- serviceを完全にCinerino化
+
 ## v30.1.1 - 2019-03-05
 
 ### Changed